IMS: Send connection event to UI for changes in phone account
Dialer needs to update RTT options based on the RTT capability
for the phone account. So send a new connection event to Dialer
whenever there is a change in the phone account for the call.
Dialer can listen to this event to see if there is any change
in RTT capability and update UI options accordingly.

CRs-Fixed: 2918278
Change-Id: If22515774b412f88d2c71f9903469a4aba1f88e3