Don't play local ring while CRS is playing
- In case MT call with video CRS, DUT shall play audio
from network, dialing any call in either of the SUB would
not be allowed, but user will be prompted with a cannot
place a call message after dial button is clicked. The failing
call update in Telecom will finish CRS and play local ring due
to not check the call is mForegroundCall or not,this is not
expected.

- To fix it, reset CRS mode only the call is mForegroundCall in
onCallStateChanged.

Change-Id: I3a9d022c47f590cad11dadac7829b9ba217b0816
CRs-Fixed: 3239138