Fix call audio switch issue for BT LE to speaker and back to BT switch
Currently when call audio is switched from BT LE headset to speaker,
first audio communication device is set as speaker and then BT audio
is disconnected. Hence onAudioLost() is not invoked when
clearLeAudioCommunicationDevice() is called. This causes
BluetoothRouteManager to wrongly assume that the BT LE headset
is in AudioConnected state. Hence when audio is switched back to BT,
BluetoothRouteManager doesn't try to connect BT audio and switch
fails. Fix this by making sure that onAudioLost() is always
invoked when clearLeAudioCommunicationDevice() is called.

Change-Id: I67663fce0e3f01e551622483e0ce2c777712e451
CRs-Fixed: 3246566