Fix ringer not stopping issue during answer of waiting call
When waiting call on other sub is answered during pseudo DSDA
mode, first the active call is hungup and then incoming call
is answered. During this time the incoming call state changes
from call waiting to ringing for a short time period and ringtone
player play and stop is requested back to back.
AyncRingtonePlayer handles play and stop requests asynchronously in
a handler thread and sometimes isPlaying() can return false if
called immediately after a play request. This is causing the
Ringer to not stop ringtone in some cases. To fix this, always
issue stop request to AsyncRingTonePlayer without checking
isPlaying().

Change-Id: I2b744c735d2f745ac71f3645918da0df0ac0154b
CRs-Fixed: 3498128