Prevent BT LE headset disconnect during switch from wired headset
When a connect request with null address is received while
BluetoothRouteManager is in connecting state, it is wrongly
treated as switching of devices and the current device is
cleared. This causes BT headset disconnect when switching
from wired headset. If null address is received in connect
request, don't treat it as device switch.

Change-Id: If7c59d073f79026305fa9224d46f74bc78933ef3
CRs-Fixed: 3494355